+++ src/routes/AuthorizationRouter.tsx
@@ -15,21 +15,30 @@ import {
 export const AuthorizationRouter = () => {
   const [params] = useSearchParams();
   const store = useAppContextStore();
+  const email = params.get('email');
   const skipLoginIntro = params.get('skipIntro') || store.whiteLabel?.skipLoginIntro;
 
   const isGame = ['metaverse-dash-run', 'candy-jam', 'cere-game-portal'].includes(store.app?.appId as string); // TODO remove after promo
+
   return (
     <Routes>
       <Route element={<Authorize />}>
-        <Route index element={isGame || skipLoginIntro ? <AuthorizeLogin /> : <AuthorizeIntro />} />
-        <Route path="close" element={<AuthorizeClose />} />
-        <Route path="redirect" element={<AuthorizeRedirect />} />
-        <Route path="intro" element={<AuthorizeIntro />} />
-        <Route path="signin" element={<AuthorizeLogin variant="signin" />} />
-        <Route path="signup" element={<AuthorizeLogin variant="signup" />} />
-        <Route path="otp" element={<AuthorizeOtp />} />
+        {email ? (
+          <Route index element={<AuthorizeOtp sendOtp />} />
+        ) : (
+          <>
+            <Route index element={isGame || skipLoginIntro ? <AuthorizeLogin /> : <AuthorizeIntro />} />
+            <Route path="intro" element={<AuthorizeIntro />} />
+            <Route path="signin" element={<AuthorizeLogin variant="signin" />} />
+            <Route path="signup" element={<AuthorizeLogin variant="signup" />} />
+            <Route path="otp" element={<AuthorizeOtp />} />
+          </>
+        )}
+
         <Route path="permissions" element={<AuthorizePermissions />} />
         <Route path="complete" element={<AuthorizeComplete />} />
+        <Route path="close" element={<AuthorizeClose />} />
+        <Route path="redirect" element={<AuthorizeRedirect />} />
       </Route>
     </Routes>
   );